+++ dashboards/tests/tests.py
@@ -1,8 +1,6 @@
-from dashboards.data_cleaning import data_formatter as dfo
 from dashboards.data_cleaning import data_filter as dfi
+from dashboards.bbrc import data_filter as dfib
 from dashboards.data_cleaning import graph_generator as gg
-from dashboards.bbrc import data_formatter as dfo_bbrc
-from dashboards.bbrc import data_filter as dfi_bbrc
 from dashboards.app import app
 from dashboards import config
 import os.path as op
@@ -36,174 +34,174 @@ def test_001_pickle_save():  # should be run first
     assert isinstance(data['longitudinal_data'], dict)
 
 
-def test_002_get_projects_details():
-    p = pickle.load(open(config.PICKLE_PATH, 'rb'))
-
-    projects = p['info']['projects']
-    project_details = dfo.Formatter().get_projects_details(projects)
-
-    assert isinstance(project_details['Number of Projects'], int)
-    assert isinstance(project_details['Projects'], dict)
-
-    assert project_details['Number of Projects'] != 0
-    assert len(project_details['Projects']) != 0
-
-
-def test_003_get_projects_details_PP():
-    p = pickle.load(open(config.PICKLE_PATH, 'rb'))
-
-    projects = p['info']['projects']
-    project = p['info']['projects'][0]
-    project_id = project['id']
-    project_details = dfo.FormatterPP(project_id).get_projects_details(projects)
-
-    assert isinstance(project_details['Owner(s)'], list)
-    assert isinstance(project_details['Collaborator(s)'], list)
-    assert isinstance(project_details['member(s)'], list)
-    assert isinstance(project_details['user(s)'], list)
-    assert isinstance(project_details['last_accessed(s)'], list)
-    assert isinstance(project_details['insert_user(s)'], str)
-    assert isinstance(project_details['insert_date'], str)
-    assert isinstance(project_details['access'], str)
-    assert isinstance(project_details['name'], str)
-    assert isinstance(project_details['last_workflow'], str)
+# def test_002_get_projects_details():
+#     p = pickle.load(open(config.PICKLE_PATH, 'rb'))
+#
+#     projects = p['info']['projects']
+#     project_details = dfo.Formatter().get_projects_details(projects)
+#
+#     assert isinstance(project_details['Number of Projects'], int)
+#     assert isinstance(project_details['Projects'], dict)
+#
+#     assert project_details['Number of Projects'] != 0
+#     assert len(project_details['Projects']) != 0
 
 
-def test_004_get_projects_details_specific():
-    p = pickle.load(open(config.PICKLE_PATH, 'rb'))
-
-    projects = p['info']['projects']
-    project_details_specific = dfo.Formatter().get_projects_details_specific(projects)
-
-    assert isinstance(project_details_specific, dict)
-    assert len(project_details_specific) != 0
-
-
-def test_005_get_subjects_details():
-    p = pickle.load(open(config.PICKLE_PATH, 'rb'))
-
-    subjects = p['info']['subjects']
-    subject_details = dfo.Formatter().get_subjects_details(subjects)
-
-    assert isinstance(subject_details['Number of Subjects'], int)
-    assert isinstance(subject_details['Subjects'], dict)
-
-    assert subject_details['Number of Subjects'] != 0
-    assert len(subject_details['Subjects']) != 0
+# def test_003_get_projects_details_PP():
+#     p = pickle.load(open(config.PICKLE_PATH, 'rb'))
+#
+#     projects = p['info']['projects']
+#     project = p['info']['projects'][0]
+#     project_id = project['id']
+#     project_details = dfo.FormatterPP(project_id).get_projects_details(projects)
+#
+#     assert isinstance(project_details['Owner(s)'], list)
+#     assert isinstance(project_details['Collaborator(s)'], list)
+#     assert isinstance(project_details['member(s)'], list)
+#     assert isinstance(project_details['user(s)'], list)
+#     assert isinstance(project_details['last_accessed(s)'], list)
+#     assert isinstance(project_details['insert_user(s)'], str)
+#     assert isinstance(project_details['insert_date'], str)
+#     assert isinstance(project_details['access'], str)
+#     assert isinstance(project_details['name'], str)
+#     assert isinstance(project_details['last_workflow'], str)
 
 
-def test_006_get_subjects_details_PP():
-    p = pickle.load(open(config.PICKLE_PATH, 'rb'))
-
-    subjects = p['info']['subjects']
-    project = p['info']['projects'][0]
-    project_id = project['id']
-    subject_details = dfo.FormatterPP(project_id).get_subjects_details(subjects)
-
-    assert isinstance(subject_details['Number of Subjects'], int)
-    assert subject_details['Number of Subjects'] != 0
+# def test_004_get_projects_details_specific():
+#     p = pickle.load(open(config.PICKLE_PATH, 'rb'))
+#
+#     projects = p['info']['projects']
+#     project_details_specific = dfo.Formatter().get_projects_details_specific(projects)
+#
+#     assert isinstance(project_details_specific, dict)
+#     assert len(project_details_specific) != 0
 
 
-def test_007_get_experiments_details():
-    p = pickle.load(open(config.PICKLE_PATH, 'rb'))
-
-    experiments = p['info']['experiments']
-    experiments_details = dfo.Formatter().get_experiments_details(experiments)
-
-    assert isinstance(experiments_details['Number of Experiments'], int)
-    assert isinstance(experiments_details['Total amount of sessions'], dict)
-
-    assert experiments_details['Number of Experiments'] != 0
-    assert len(experiments_details['Total amount of sessions']) != 0
-
-
-def test_008_get_experiments_details_PP():
-    p = pickle.load(open(config.PICKLE_PATH, 'rb'))
-
-    experiments = p['info']['experiments']
-    project = p['info']['projects'][0]
-    project_id = project['id']
-    experiments_details = dfo.FormatterPP(project_id).get_experiments_details(experiments)
-
-    assert isinstance(experiments_details['Number of Experiments'], int)
-    assert isinstance(experiments_details['Total amount of sessions'], dict)
-
-    assert experiments_details['Number of Experiments'] != 0
-    assert len(experiments_details['Total amount of sessions']) != 0
+# def test_005_get_subjects_details():
+#     p = pickle.load(open(config.PICKLE_PATH, 'rb'))
+#
+#     subjects = p['info']['subjects']
+#     subject_details = dfo.Formatter().get_subjects_details(subjects)
+#
+#     assert isinstance(subject_details['Number of Subjects'], int)
+#     assert isinstance(subject_details['Subjects'], dict)
+#
+#     assert subject_details['Number of Subjects'] != 0
+#     assert len(subject_details['Subjects']) != 0
 
 
-def test_009_get_scans_details():
-    p = pickle.load(open(config.PICKLE_PATH, 'rb'))
-    # bbrc_resources = p['extra_resources']
-    resources, bbrc_resources = [], []
-    for e in p['resources']:
-        if len(e) == 4:
-            resources.append(e)
-        elif len(e) > 4:
-            bbrc_resources.append(e)
-
-    scans = p['info']['scans']
-    scans_details = dfo.Formatter().get_scans_details(scans)
-
-    assert isinstance(scans_details['Number of Scans'], int)
-    assert isinstance(scans_details['Scan quality'], dict)
-    assert isinstance(scans_details['Scan Types'], dict)
-
-    assert scans_details['Number of Scans'] != 0
-    assert len(scans_details['Scan quality']) != 0
-    assert len(scans_details['Scan Types']) != 0
+# def test_006_get_subjects_details_PP():
+#     p = pickle.load(open(config.PICKLE_PATH, 'rb'))
+#
+#     subjects = p['info']['subjects']
+#     project = p['info']['projects'][0]
+#     project_id = project['id']
+#     subject_details = dfo.FormatterPP(project_id).get_subjects_details(subjects)
+#
+#     assert isinstance(subject_details['Number of Subjects'], int)
+#     assert subject_details['Number of Subjects'] != 0
 
 
-def test_010_get_scans_details_PP():
-    p = pickle.load(open(config.PICKLE_PATH, 'rb'))
-
-    scans = p['info']['scans']
-    project = p['info']['projects'][0]
-    project_id = project['id']
-    scans_details = dfo.FormatterPP(project_id).get_scans_details(scans)
-
-    assert isinstance(scans_details['Number of Scans'], int)
-    assert isinstance(scans_details['Scan quality'], dict)
-    assert isinstance(scans_details['Scan Types'], dict)
-
-    assert scans_details['Number of Scans'] != 0
-    assert len(scans_details['Scan quality']) != 0
-    assert len(scans_details['Scan Types']) != 0
+# def test_007_get_experiments_details():
+#     p = pickle.load(open(config.PICKLE_PATH, 'rb'))
+#
+#     experiments = p['info']['experiments']
+#     experiments_details = dfo.Formatter().get_experiments_details(experiments)
+#
+#     assert isinstance(experiments_details['Number of Experiments'], int)
+#     assert isinstance(experiments_details['Total amount of sessions'], dict)
+#
+#     assert experiments_details['Number of Experiments'] != 0
+#     assert len(experiments_details['Total amount of sessions']) != 0
 
 
-def test_011_get_resources_details():
-    p = pickle.load(open(config.PICKLE_PATH, 'rb'))
-    # bbrc_resources = p['extra_resources']
-    resources, bbrc_resources = [], []
-    for e in p['resources']:
-        if len(e) == 4:
-            resources.append(e)
-        elif len(e) > 4:
-            bbrc_resources.append(e)
+# def test_008_get_experiments_details_PP():
+#     p = pickle.load(open(config.PICKLE_PATH, 'rb'))
+#
+#     experiments = p['info']['experiments']
+#     project = p['info']['projects'][0]
+#     project_id = project['id']
+#     experiments_details = dfo.FormatterPP(project_id).get_experiments_details(experiments)
+#
+#     assert isinstance(experiments_details['Number of Experiments'], int)
+#     assert isinstance(experiments_details['Total amount of sessions'], dict)
+#
+#     assert experiments_details['Number of Experiments'] != 0
+#     assert len(experiments_details['Total amount of sessions']) != 0
 
-    resource_details = dfo.Formatter().get_resources_details(resources)
 
-    assert isinstance(resource_details, dict)
-    assert len(resource_details) != 0
+# def test_009_get_scans_details():
+#     p = pickle.load(open(config.PICKLE_PATH, 'rb'))
+#     # bbrc_resources = p['extra_resources']
+#     resources, bbrc_resources = [], []
+#     for e in p['resources']:
+#         if len(e) == 4:
+#             resources.append(e)
+#         elif len(e) > 4:
+#             bbrc_resources.append(e)
+#
+#     scans = p['info']['scans']
+#     scans_details = dfo.Formatter().get_scans_details(scans)
+#
+#     assert isinstance(scans_details['Number of Scans'], int)
+#     assert isinstance(scans_details['Scan quality'], dict)
+#     assert isinstance(scans_details['Scan Types'], dict)
+#
+#     assert scans_details['Number of Scans'] != 0
+#     assert len(scans_details['Scan quality']) != 0
+#     assert len(scans_details['Scan Types']) != 0
 
 
-def test_012_get_resources_details_PP():
+# def test_010_get_scans_details_PP():
+#     p = pickle.load(open(config.PICKLE_PATH, 'rb'))
+#
+#     scans = p['info']['scans']
+#     project = p['info']['projects'][0]
+#     project_id = project['id']
+#     scans_details = dfo.FormatterPP(project_id).get_scans_details(scans)
+#
+#     assert isinstance(scans_details['Number of Scans'], int)
+#     assert isinstance(scans_details['Scan quality'], dict)
+#     assert isinstance(scans_details['Scan Types'], dict)
+#
+#     assert scans_details['Number of Scans'] != 0
+#     assert len(scans_details['Scan quality']) != 0
+#     assert len(scans_details['Scan Types']) != 0
 
-    p = pickle.load(open(config.PICKLE_PATH, 'rb'))
-    # bbrc_resources = p['extra_resources']
-    resources, bbrc_resources = [], []
-    for e in p['resources']:
-        if len(e) == 4:
-            resources.append(e)
-        elif len(e) > 4:
-            bbrc_resources.append(e)
+#
+# def test_011_get_resources_details():
+#     p = pickle.load(open(config.PICKLE_PATH, 'rb'))
+#     # bbrc_resources = p['extra_resources']
+#     resources, bbrc_resources = [], []
+#     for e in p['resources']:
+#         if len(e) == 4:
+#             resources.append(e)
+#         elif len(e) > 4:
+#             bbrc_resources.append(e)
+#
+#     resource_details = dfo.Formatter().get_resources_details(resources)
+#
+#     assert isinstance(resource_details, dict)
+#     assert len(resource_details) != 0
 
-    project = p['info']['projects'][0]
-    project_id = project['id']
-    resource_details = dfo.FormatterPP(project_id).get_resources_details(resources)
 
-    assert isinstance(resource_details, dict)
-    assert len(resource_details) != 0
+# def test_012_get_resources_details_PP():
+#
+#     p = pickle.load(open(config.PICKLE_PATH, 'rb'))
+#     # bbrc_resources = p['extra_resources']
+#     resources, bbrc_resources = [], []
+#     for e in p['resources']:
+#         if len(e) == 4:
+#             resources.append(e)
+#         elif len(e) > 4:
+#             bbrc_resources.append(e)
+#
+#     project = p['info']['projects'][0]
+#     project_id = project['id']
+#     resource_details = dfo.FormatterPP(project_id).get_resources_details(resources)
+#
+#     assert isinstance(resource_details, dict)
+#     assert len(resource_details) != 0
 
 
 def test_013_get_get_longitudinal_data():
@@ -215,57 +213,57 @@ def test_013_get_get_longitudinal_data():
     assert len(long_data) != 0
 
 
-def test_014_get_bbrc_resource_details():
-    p = pickle.load(open(config.PICKLE_PATH, 'rb'))
-    # bbrc_resources = p['extra_resources']
-    resources, bbrc_resources = [], []
-    for e in p['resources']:
-        if len(e) == 4:
-            resources.append(e)
-        elif len(e) > 4:
-            bbrc_resources.append(e)
-
-    resource_bbrc_details = dfo_bbrc.Formatter().get_resource_details(bbrc_resources)
-
-    assert isinstance(resource_bbrc_details, dict)
-    assert len(resource_bbrc_details) != 0
-
-
-def test_015_diff_dates():
-    p = pickle.load(open(config.PICKLE_PATH, 'rb'))
-    # bbrc_resources = p['extra_resources']
-    resources, bbrc_resources = [], []
-    for e in p['resources']:
-        if len(e) == 4:
-            resources.append(e)
-        elif len(e) > 4:
-            bbrc_resources.append(e)
-    #experiments = p['info']['experiments']
-    project = p['info']['projects'][0]
-    project_id = project['id']
-    dict_diff_dates = dfo_bbrc.Formatter().diff_dates(bbrc_resources, project_id)
-
-    assert isinstance(dict_diff_dates, dict)
-    assert len(dict_diff_dates) != 0
+# def test_014_get_bbrc_resource_details():
+#     p = pickle.load(open(config.PICKLE_PATH, 'rb'))
+#     # bbrc_resources = p['extra_resources']
+#     resources, bbrc_resources = [], []
+#     for e in p['resources']:
+#         if len(e) == 4:
+#             resources.append(e)
+#         elif len(e) > 4:
+#             bbrc_resources.append(e)
+#
+#     resource_bbrc_details = dfo_bbrc.Formatter().get_resource_details(bbrc_resources)
+#
+#     assert isinstance(resource_bbrc_details, dict)
+#     assert len(resource_bbrc_details) != 0
 
 
-def test_016_generate_test_grid_bbrc():
+# def test_015_diff_dates():
+#     p = pickle.load(open(config.PICKLE_PATH, 'rb'))
+#     # bbrc_resources = p['extra_resources']
+#     resources, bbrc_resources = [], []
+#     for e in p['resources']:
+#         if len(e) == 4:
+#             resources.append(e)
+#         elif len(e) > 4:
+#             bbrc_resources.append(e)
+#     #experiments = p['info']['experiments']
+#     project = p['info']['projects'][0]
+#     project_id = project['id']
+#     dict_diff_dates = dfo_bbrc.Formatter().diff_dates(bbrc_resources, project_id)
+#
+#     assert isinstance(dict_diff_dates, dict)
+#     assert len(dict_diff_dates) != 0
 
-    p = pickle.load(open(config.PICKLE_PATH, 'rb'))
-    # bbrc_resources = p['extra_resources']
-    resources, bbrc_resources = [], []
-    for e in p['resources']:
-        if len(e) == 4:
-            resources.append(e)
-        elif len(e) > 4:
-            bbrc_resources.append(e)
 
-    project = p['info']['projects'][0]
-    project_id = project['id']
-    test_grid = dfo_bbrc.Formatter().generate_test_grid_bbrc(bbrc_resources, project_id)
-
-    assert isinstance(test_grid, list)
-    assert len(test_grid) != 0
+# def test_016_generate_test_grid_bbrc():
+#
+#     p = pickle.load(open(config.PICKLE_PATH, 'rb'))
+#     # bbrc_resources = p['extra_resources']
+#     resources, bbrc_resources = [], []
+#     for e in p['resources']:
+#         if len(e) == 4:
+#             resources.append(e)
+#         elif len(e) > 4:
+#             bbrc_resources.append(e)
+#
+#     project = p['info']['projects'][0]
+#     project_id = project['id']
+#     test_grid = dfo_bbrc.Formatter().generate_test_grid_bbrc(bbrc_resources, project_id)
+#
+#     assert isinstance(test_grid, list)
+#     assert len(test_grid) != 0
 
 
 def test_017_get_project_list():
@@ -335,7 +333,7 @@ def test_022_reorder_graphs_bbrc():
         elif len(e) > 4:
             bbrc_resources.append(e)
 
-    filtered = dfi_bbrc.DataFilter('admin', [], bbrc_resources)
+    filtered = dfib.BBRCDataFilter('admin', [], bbrc_resources)
     ordered_graphs = filtered.reorder_graphs()
 
     assert isinstance(ordered_graphs, dict)
@@ -356,7 +354,7 @@ def test_023_reorder_graphs_bbrc_PP():
     project = p['info']['projects'][0]
     project_id = project['id']
     role = 'admin'
-    filtered = dfi_bbrc.DataFilterPP(experiments, project_id,
+    filtered = dfib.DataFilterPP(experiments, project_id,
                                      role, {role: [project_id]}, bbrc_resources)
     ordered_graphs = filtered.reorder_graphs_pp()
     assert isinstance(ordered_graphs, dict)
@@ -413,7 +411,7 @@ def test_028_stats():
             sess['server'] = 'https://devxnat.barcelonabeta.org'
             sess['role'] = 'guest'
             sess['username'] = 'testUser'
-            sess['project_visible'] = []
+            sess['projects'] = []
 
     response_get_stats = c.get('dashboard/stats/').status_code
     assert response_get_stats == 200
